+++ src/main/java/com/hotsix/iAmNotAlone/login/config/auth/PrincipalDetails.java
@@ -1,6 +1,6 @@
 package com.hotsix.iAmNotAlone.login.config.auth;
 
-import com.hotsix.iAmNotAlone.domain.Member;
+import com.hotsix.iAmNotAlone.signup.domain.Member;
 import lombok.Getter;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
