+++ src/main/java/com/hotsix/iAmNotAlone/login/config/service/JwtService.java
@@ -3,8 +3,7 @@ package com.hotsix.iAmNotAlone.login.config.service;
 import com.auth0.jwt.JWT;
 import com.auth0.jwt.algorithms.Algorithm;
 import com.hotsix.iAmNotAlone.login.config.auth.PrincipalDetails;
-import com.hotsix.iAmNotAlone.login.config.jwt.JwtProperties;
-import com.hotsix.iAmNotAlone.repository.MemberRepository;
+import com.hotsix.iAmNotAlone.signup.repository.MemberRepository;
 import lombok.AccessLevel;
 import lombok.RequiredArgsConstructor;
 import lombok.Setter;
