+++ src/main/resources/application.yml
@@ -4,10 +4,10 @@ spring:
       enabled: true
       path: /h2-console
   datasource:
-    driver-class-name: com.mysql.cj.jdbc.Driver
-    url: jdbc:mysql://localhost:3306/orders?serverTimezone=Asia/Seoul
-    username: root
-    password: zerobase
+    driver-class-name: org.h2.Driver
+    url: jdbc:h2:mem:test
+    username: sa
+    password:
 
   jpa:
     hibernate:
